Introduce a proper test utility module
Heaps of our tests run multiple times, because test_tc and test_xdp use
the test module, and with that pull in the tests it defines. Introduce a
proper module with common functionality to fix this.
